rumit

+++ resources/views/mapel/index.blade.php
@@ -1,23 +1,7 @@
-<!DOCTYPE html>
-<html lang="en">
+@extends('layout.mainlayout')
+@section('title', 'Halaman Mapel')
 
-<head>
-    <meta charset="UTF-8">
-    <meta content="width=device-width, initial-scale=1.0" name="viewport">
-    <meta content="ie=edge" http-equiv="X-UA-Compatible">
-    <title>Mata Pelajaran</title>
-    {{-- <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.1/dist/jquery.min.js"></script> --}}
-    @vite('resources/css/app.css')
-    @vite('resources/js/app.js')
-</head>
-
-<body>
-
-
-    @include('layout.navbar')
-
-
-
+@section('content')
     <div class="overflow-x-auto relative shadow-md sm:rounded-lg pt-10">
         <div class="flex justify-between items-center pb-4">
             <div>
@@ -69,8 +53,7 @@
                 @foreach ($mapel as $g)
                     <tr
                         class="bg-white border-b dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-600">
-                        <th class="py-4 px-6 font-medium text-gray-900 whitespace-nowrap dark:text-white"
-                            scope="row">
+                        <th class="py-4 px-6 font-medium text-gray-900 whitespace-nowrap dark:text-white" scope="row">
                             {{ $g->MapelId }}
                         </th>
                         <td class="py-4 px-6">
@@ -80,8 +63,7 @@
                             <a class="font-medium text-blue-600 dark:text-blue-500 hover:underline"
                                 data-modal-toggle="Edit{{ $g->MapelId }}" href="#" type="button">Edit</a>
                             <a class="font-medium text-red-600 dark:text-red-500 hover:underline"
-                                href="/mapel/{{ $g->MapelId }}/hapus"
-                                onclick="return confirm('Hapus Data ?')">Remove</a>
+                                href="/mapel/{{ $g->MapelId }}/hapus" onclick="return confirm('Hapus Data ?')">Remove</a>
                             @include('mapel.edit')
                         </td>
                     </tr>
@@ -90,6 +72,4 @@
         </table>
 
         @include('mapel.create')
-</body>
-
-</html>
+    @endsection
